refactor(agent-task): 스킬 바인딩 해석을 skill-block 으로 이동 (파일 크기 가드)
AgentTaskService 가 정확히 600줄(가드 상한)이라 이번 기능의 5줄 추가만으로
CI Gate 3 가 605줄로 실패했다. 크램 대신 스킬 관련 로직을 제자리로 옮긴다 —
바인딩 조회·allowedSkills 범위 제한·도구 머지를 resolveSkillToolBindings 로
추출(이미 스킬을 다루는 agent-task/skill-block). 589줄로 여유 확보.

동작 변화 없음. 라이브 재검증(브라우저 에이전트 작업): 카탈로그에서
"데이터 분석가 전문 스킬" 자율 선택 → 로드 → 보고 정상.
